+++ CHANGELOG.md
@@ -21,6 +21,11 @@ For each release, changes are grouped under these headings:
 
 The following changes have been completed, and will be included in the next tagged release.
 
+### New
+
+* Added some helpers for working with `catch` values:
+  - added `extractReasonFromCaught()`
+
 ### Tools
 
 * Upgraded the `scripts` section in `package.json` to our latest standard.